+++ src/main/java/kr/elroy/aigoya/order/Order.java
@@ -7,7 +7,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
-import kr.elroy.aigoya.product.OrderProduct;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
