+++ src/main/java/kr/elroy/aigoya/order/OrderProduct.java
@@ -1,4 +1,4 @@
-package kr.elroy.aigoya.product;
+package kr.elroy.aigoya.order;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,7 +7,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import kr.elroy.aigoya.order.Order;
+import kr.elroy.aigoya.product.Product;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
